New Final Topic Model (V2)
Created a new LDA topic model (second version) by adding more stopwords found in previous iterations of the LDA algorithm. We have decided to include top 1% of the words in the junk topics as stop words in this new iteration. This update contains a new notebook, visualizations and saved LDA model files.