Generated complete topic data for V2 LDA 18 topic model
Previously, we decided to work with the 4 topic model, but it turned out that it might be a bit too simplistic. We have plans to revert to the 18 topic model from now on, so that is why this data was generated now and not previously.

Extra: Saved corpora, id2word dictionary and added instructions for loading and saving